+++ components/app-connector-validation-proxy/cmd/appconnectorvalidationproxy/appconnectorvalidationproxy.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/kyma-project/kyma/components/application-connector-dispatcher/internal/validationproxy"
+	"github.com/kyma-project/kyma/components/app-connector-validation-proxy/internal/validationproxy"
 	log "github.com/sirupsen/logrus"
 )
