+++ components/app-connector-validation-proxy/internal/validationproxy/mocks/ApplicationManager.go
