--- components/application-connector-dispatcher/.gitignore
@@ -1,3 +1,0 @@
-/.idea
-*.iml
-/validationproxy
