--- components/application-connector-dispatcher/Dockerfile.validation-proxy
@@ -1,15 +1,0 @@
-FROM golang:1.12-alpine as builder
-
-ARG DOCK_PKG_DIR=/go/src/github.com/kyma-project/kyma/components/application-connector-dispatcher
-
-WORKDIR $DOCK_PKG_DIR
-COPY . $DOCK_PKG_DIR
-
-RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o validationproxy ./cmd/validationproxy
-
-FROM scratch
-LABEL source=git@github.com:kyma-project/kyma.git
-
-COPY --from=builder /go/src/github.com/kyma-project/kyma/components/application-connector-dispatcher .
-
-CMD ["/validationproxy"]
